fixed navigation and token issues
