+++ navigation/MainDrawerNavigator.js
@@ -5,6 +5,7 @@ import { StackNavigator, DrawerNavigator } from 'react-navigation';
 import { Colors, Layout } from '../constants/';
 
 import LoginScreen from '../screens/LoginScreen';
+import CountryScreen from '../screens/CountryScreen';
 import HomeScreen from '../screens/HomeScreen';
 import FavoriteScreen from '../screens/FavoriteScreen';
 import NotificationScreen from '../screens/NotificationScreen';
@@ -27,6 +28,7 @@ import DrawerContainer from '../screens/DrawerContainer/';
 // drawer stack
 const DrawerStack = DrawerNavigator(
   {
+    countryScreen: { screen:CountryScreen },
     loginscreen: { screen:LoginScreen },
     homeScreen: { screen: HomeScreen },
     favoriteScreen: { screen: FavoriteScreen },
