+++ screens/LoginScreen/index.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Image, View, ImageBackground } from 'react-native';
+import { Image, View, ImageBackground,BackHandler } from 'react-native';
 //import { Mutation } from 'react-apollo';
 //import gql from 'graphql-tag';
 import { Container, Content, Left, Text, Button } from 'native-base';
@@ -27,33 +27,12 @@ export default class LoginScreen extends React.Component {
     }
   }
 
-  // Check weather user is login or not
   componentWillMount = async () => {
-
-    let jwtt = '';
-    jwtt = Expo.SecureStore.getItemAsync('JWTToken').then();
-    console.log("Chat Log : " + jwtt);
-
-    if(jwtt == '' || jwtt == null || jwtt.length == 0)
-    {
-      let arriveFrom = '';
-      arriveFrom = Expo.SecureStore.getItemAsync('ArrivedFrom').then();
-      this.setState({ArriverFrom: arriveFrom})
-      console.log("Chat Log : " + this.state.ArriverFrom);
-
-      if(this.state.ArriverFrom == 'ChatListScreen')
-      {
-        this.props.navigation.navigate('ChatListScreen');
-      }
-      if(this.state.ArriverFrom == 'ProfileScreen')
-      {
-        this.props.navigation.navigate('ProfileScreen');
-      }
-      if(this.state.ArriverFrom == 'CreateNewItemScreen')
-      {
-        this.props.navigation.navigate('CreateNewItemScreen');
-      }
-    }
+    var that = this
+    BackHandler.addEventListener('hardwareBackPress', function() {
+      that.props.navigation.navigate('homeScreen')
+      return true;
+    });
 
   }
